requirements.txt
requests==2.32.3
auto-py-to-exe==2.44.1
python-dotenv==1.0.1